rewrote argument handling
Previously, the argument handling was based on a literal interpretation
of the spec's description of 0x1b - namely, that its behavior varied
depending on whether it was in a or b. After reading the spec more closely
(especially in areas surrounding comparison and special operations), as well
as comparing the behavior of other emulators, I decided to switch to a model
based on separate read and write operations.

Instruction and argument parsing was also moved, so that instead of working
with bits in the instruction loop directly, instructions and arguments are
parsed into a data structure. (This is a simple struct of roughly 10 bytes.)